fix edit & delete proudct routes

+++ bsound-server/routes/product.routes.js
@@ -3,13 +3,14 @@ const User = require("../models/User.model");
 const Product = require('../models/Product.model');
 const fileUploader = require("../config/cloudinary.config");
 const { isAuthenticated } = require("../middleware/jwt.middleware");
+const mongoose = require('mongoose');
 
 // GET - get all the products availables
-router.get("/", (req, res) => {
+router.get("/products", (req, res, next) => {
     Product.find()
-      .populate("owner", "name")
-      .then((products) => {
-        res.json(products);
+      .populate("owner", "name") // from user
+      .then((allProducts) => {
+        res.json(allProducts);
       })
       .catch((error) => {
         res.status(500).json({ error: "Error to get all the products." });
@@ -17,15 +18,20 @@ router.get("/", (req, res) => {
   });
 
 // GET - get the product details
-router.get("/:productId", (req, res) => {
+router.get("/products/:productId", (req, res, next) => {
     const { productId } = req.params;
+
+    if(!mongoose.Types.ObjectId.isValid(productId)) {
+      res.status(400).json({ message: 'Specified id is not valid' });
+      return;
+  }
   
     Product.findById(productId)
       .populate("owner", "name")
       .then((product) => {
-        if (!product) {
-          return res.status(404).json({ error: "Product not found." });
-        }
+        // if (!product) {
+        //   return res.status(404).json({ error: "Product not found." });
+        // }
         res.json(product);
       })
       .catch((error) => {
@@ -34,38 +40,70 @@ router.get("/:productId", (req, res) => {
   });
 
 // POST - cretae a new product
-router.post("/", isAuthenticated, fileUploader.array("mediaUrl", 5), (req, res) => {
-    const { productName, category, description, availableDays, mediaUrl, pricePerDay, location, contactDetails } = req.body;
-    const owner = req.user.id;
-    const mediaUrls = req.files.map((file) => file.path);
-  
-    const newProduct = new Product({
+router.post("/products", isAuthenticated, fileUploader.single("mediaUrl"), (req, res, next) => {
+
+//   router.post("/products", isAuthenticated, fileUploader.array("mediaUrl", 5), (req, res, next) => {
+
+  // const images = req.files.map((file) => file.path);
+
+    let { productName, category, description,  availability: { 
+      startDate,
+      endDate,
+    }, mediaUrl, pricePerDay, location, contactDetails } = req.body;
+    // const owner = req.user;
+    let owner = req.payload._id;
+    
+    // if (Array.isArray(req.files) && req.files.length > 0) {
+    // const mediaUrls = req.files.map((file) => file.path);
+
+    let newProduct = new Product({
       productName,
       category,
       description,
-      availableDays,
-      mediaUrl: mediaUrls,
+      // availableDays,
+      availability: { 
+        startDate,
+        endDate,
+      },
       pricePerDay,
       location,
       contactDetails,
       owner,
+      mediaUrl
+     // mediaUrl: mediaUrls
     });
+
+    mediaUrl = req.file ? req.file.path : undefined;
   
+    console.log(newProduct)
     newProduct
       .save()
       .then((product) => {
         res.status(201).json(product);
       })
       .catch((error) => {
+        console.log(error)
         res.status(400).json({ error: "Product cannot be created." });
       });
+
+    // } else {
+    //   res.status(400).json({ error: "Invalid mediaUrl. Please upload at least one file." });
+    // }
   });
 
 // PUT - edit an existing product
-router.put("/:productId", isAuthenticated, fileUploader.array("mediaUrl", 5), (req, res) => {
+router.put("/products/:productId", isAuthenticated, fileUploader.single("mediaUrl"), (req, res, next) => {
     const { productId } = req.params;
-    const { productName, category, description, availableDays, mediaUrl, pricePerDay, location, contactDetails } = req.body;
-    const mediaUrls = req.files.map((file) => file.path); 
+    const { productName, category, description,  availability: { 
+      startDate,
+      endDate,
+    } , mediaUrl, pricePerDay, location, contactDetails } = req.body;
+    // const mediaUrls = req.files.map((file) => file.path); 
+
+    if(!mongoose.Types.ObjectId.isValid(productId)) {
+      res.status(400).json({ message: 'Specified id is not valid' });
+      return;
+  }
   
     Product.findByIdAndUpdate(
       productId,
@@ -73,8 +111,12 @@ router.put("/:productId", isAuthenticated, fileUploader.array("mediaUrl", 5), (r
         productName,
         category,
         description,
-        availableDays,
-        mediaUrl: mediaUrls,
+        // availableDays,
+        availability: { 
+          startDate,
+          endDate,
+        },
+        mediaUrl,
         pricePerDay,
         location,
         contactDetails,
@@ -82,9 +124,9 @@ router.put("/:productId", isAuthenticated, fileUploader.array("mediaUrl", 5), (r
       { new: true }
     )
       .then((updatedProduct) => {
-        if (!updatedProduct) {
-          return res.status(404).json({ error: "Product not found." });
-        }
+        // if (!updatedProduct) {
+        //   return res.status(404).json({ error: "Product not found." });
+        // }
         res.json(updatedProduct);
       })
       .catch((error) => {
@@ -93,19 +135,47 @@ router.put("/:productId", isAuthenticated, fileUploader.array("mediaUrl", 5), (r
   });
 
 // DELETE 
-router.delete("/:productId", isAuthenticated, (req, res) => {
+// router.delete("products/:productId", isAuthenticated, (req, res, next) => {
+//     const { productId } = req.params;
+
+//     if(!mongoose.Types.ObjectId.isValid(productId)) {
+//       res.status(400).json({ message: 'Specified id is not valid' });
+//       return;
+//   }
+  
+//     Product.findByIdAndRemove(productId)
+//       .then(() => {
+//         if (!deletedProduct) {
+//           return res.status(404).json({ error: "Product not found." });
+//         }
+//         res.json({ message: "Product is removed successfully." });
+//       })
+//       .catch((error) => {
+//         res.status(400).json({ error: "Product cannot be removed." });
+//       });
+//   });
+
+  router.delete("/products/:productId", isAuthenticated, (req, res, next) => {
     const { productId } = req.params;
+  
+    if (!mongoose.Types.ObjectId.isValid(productId)) {
+      res.status(400).json({ message: 'Specified id is not valid' });
+      return;
+    }
   
     Product.findByIdAndRemove(productId)
       .then((deletedProduct) => {
         if (!deletedProduct) {
           return res.status(404).json({ error: "Product not found." });
         }
-        res.json({ message: "Product removed successfully." });
+        res.json({ message: "Product is removed successfully." });
       })
       .catch((error) => {
         res.status(400).json({ error: "Product cannot be removed." });
       });
   });
+  
+  
+  
   
   module.exports = router;
